rename to inputs-common
